+++ catalog/app/utils/workflows.js
@@ -1,5 +1,6 @@
 import * as R from 'ramda'
 
+import * as s3paths from 'utils/s3paths'
 import yaml from 'utils/yaml'
 
 export const notAvaliable = Symbol('not available')
@@ -13,8 +14,22 @@ function getNoWorkflow(data, hasConfig) {
   }
 }
 
+const COPY_DATA_DEFAULT = true
+
 export const emptyConfig = {
+  copyData: COPY_DATA_DEFAULT,
+  successors: [],
   workflows: [getNoWorkflow({}, false)],
+}
+
+export function shouldSuccessorCopyData(workflowsConfig, bucket) {
+  if (!workflowsConfig.successors || !workflowsConfig.successors.length)
+    return COPY_DATA_DEFAULT
+
+  const successor = R.find(R.propEq('slug', bucket), workflowsConfig.successors)
+  if (!successor) return COPY_DATA_DEFAULT
+
+  return successor.copyData
 }
 
 function parseSchema(schemaSlug, schemas) {
@@ -33,6 +48,13 @@ function parseWorkflow(workflowSlug, workflow, data) {
   }
 }
 
+const parseSuccessor = (url, successor) => ({
+  copyData: successor.copy_data === undefined ? COPY_DATA_DEFAULT : successor.copy_data,
+  name: successor.title,
+  slug: s3paths.parseS3Url(url).bucket,
+  url,
+})
+
 export function parse(workflowsYaml) {
   const data = yaml(workflowsYaml)
   if (!data) return emptyConfig
@@ -48,7 +70,11 @@ export function parse(workflowsYaml) {
   const noWorkflow =
     data.is_workflow_required === false ? getNoWorkflow(data, true) : null
 
+  const successors = data.successors || {}
   return {
+    successors: Object.entries(successors).map(([url, successor]) =>
+      parseSuccessor(url, successor),
+    ),
     workflows: noWorkflow ? [noWorkflow, ...workflowsList] : workflowsList,
   }
 }